+++ lua/core/autocmds.lua
@@ -1,11 +1,9 @@
 -- Auto commands
 
-local api = vim.api
-
 -- Generic
-api.nvim_create_augroup("Generic", { clear = true })
+vim.api.nvim_create_augroup("Generic", { clear = true })
 -- Highlights yanked text
-api.nvim_create_autocmd("TextYankPost", {
+vim.api.nvim_create_autocmd("TextYankPost", {
    group = "Generic",
    pattern = { "*" },
    callback = function()
@@ -15,7 +13,7 @@ api.nvim_create_autocmd("TextYankPost", {
 })
 -- SignColumn always on, and of length 1, so new events in the SignColumn do
 -- not push the text to the right
-api.nvim_create_autocmd("VimEnter", {
+vim.api.nvim_create_autocmd("VimEnter", {
    group = "Generic",
    pattern = { "*" },
    callback = function()
@@ -25,7 +23,7 @@ api.nvim_create_autocmd("VimEnter", {
    desc = "SignColumn always on, length 1",
 })
 -- Set SignColumn color to background color, aesthetically nicer
-api.nvim_create_autocmd({ "ColorScheme", "VimEnter" }, {
+vim.api.nvim_create_autocmd({ "ColorScheme", "VimEnter" }, {
    group = "Generic",
    pattern = { "*" },
    command = "hi! link SignColumn Normal",
@@ -62,10 +60,24 @@ vim.api.nvim_create_autocmd("WinLeave", {
    desc = "Set cursorline/cursorlineopt in inactive buffer",
 })
 
+-- Restore last cursor position and centre the screen when reopening a file
+vim.api.nvim_create_augroup("RestoreCursor", { clear = true })
+vim.api.nvim_create_autocmd("BufReadPost", {
+   group = "RestoreCursor",
+   desc = "Jump to last edit position in file",
+   callback = function()
+      local last_line = vim.fn.line([['"]])
+      local total_lines = vim.fn.line("$")
+      if last_line > 1 and last_line <= total_lines then
+         vim.cmd('normal! g`"zz')
+      end
+   end,
+})
+
 -- Terminal
-api.nvim_create_augroup("Terminal", { clear = true })
+vim.api.nvim_create_augroup("Terminal", { clear = true })
 -- Disable spell checking and line numbering
-api.nvim_create_autocmd("TermOpen", {
+vim.api.nvim_create_autocmd("TermOpen", {
    group = "Terminal",
    pattern = { "*" },
    callback = function()
@@ -73,34 +85,38 @@ api.nvim_create_autocmd("TermOpen", {
       vim.opt_local.number = false
       vim.opt_local.relativenumber = false
    end,
-   desc = "Disable spell checking and line numbering in Term windows",
+   desc = "Disable spell checking and line numbering in terminal windows",
 })
--- Enter Term windows in Insert mode
-api.nvim_create_autocmd("TermOpen", {
+-- Enter terminal windows in Insert mode
+vim.api.nvim_create_autocmd("TermOpen", {
    group = "Terminal",
    pattern = { "*" },
-   callback = function(opts)
+   callback = function()
+      -- Prevent nvim-dap-ui from switching to Insert mode, see
       -- https://github.com/mfussenegger/nvim-dap/issues/439#issuecomment-1380787919
-      if opts.file:match("dap%-terminal") then
+      -- opts.file:match('dap%-terminal') doesn't work anymore
+      local buf_no = vim.api.nvim_get_current_buf()
+      local buf_ft = vim.api.nvim_get_option_value("filetype", { buf = buf_no })
+      if buf_ft == "dapui_console" then
          return
       end
       vim.cmd("startinsert")
       vim.opt_local.number = false
    end,
-   desc = "Enter Term windows in Insert mode",
+   desc = "Enter terminal windows in Insert mode",
 })
--- Exit Term windows without pressing any key
--- api.nvim_create_autocmd("TermClose", {
+-- Exit terminal windows without pressing any key
+-- vim.api.nvim_create_autocmd("TermClose", {
 --    group = "Terminal",
 --    pattern = { "*" },
 --    command = "call feedkeys('q')",
---    desc = "Exit Term windows without pressing any key",
+--    desc = "Exit terminal windows without pressing any key",
 -- })
 
 -- Quickfix lists
-api.nvim_create_augroup("Quickfix", { clear = true })
+vim.api.nvim_create_augroup("Quickfix", { clear = true })
 -- Disable spell checking
-api.nvim_create_autocmd("FileType", {
+vim.api.nvim_create_autocmd("FileType", {
    group = "Quickfix",
    pattern = { "qf" },
    callback = function()
@@ -130,18 +146,17 @@ vim.api.nvim_create_autocmd({
 })
 
 -- Restore the state of transparent background on ColorScheme event
--- Not enabled in Neovide sessions
+-- This auto command is not enabled in Neovide sessions
 if not vim.g.neovide then
+   vim.api.nvim_create_augroup("TransparentBackground", { clear = true })
    vim.api.nvim_create_autocmd("ColorScheme", {
+      group = "TransparentBackground",
       callback = function()
-         local ui = require("core.ui")
          local state = require("core.state")
-         -- TODO: Eventually remove the line below.
-         -- Why blocking for 1 ms? see lua/plugins/nvim-colorizer.lua
-         vim.wait(1) -- blocks for 1 ms
-         ui.set_background_transparency(
-            state.background_transparency_enabled_at_startup
-         )
+         local ui = require("core.ui")
+         if state.background_transparency_enabled_at_startup then
+            ui.toggle_background_transparency()
+         end
       end,
    })
 end